requirements.txt
Django==3.1
Pillow
psycopg2-binary==2.8.5
djangorestframework==3.11
python-decouple==3.4
drf-yasg
django-filter